+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix token deletion not hiding the tooltip
 - Changing actor size didn't save it on its token(s)
 - Resource values were no longer able to be referenced in formulas (#885)
+- Deleting an actor or item with an open sheet caused the sheet to be uncloseable
 
 ### Changelog